--- a/.travis.yml
+++ b/.travis.yml
@@ -5,35 +5,17 @@ os:
   - osx
 
 go:
-  - 1.2.x
-  - 1.3.x
-  - 1.4
-  - 1.5.x
-  - 1.6.x
-  - 1.7.x
-  - 1.8.x
-  - 1.9.x
   - 1.10.x
   - 1.11.x
   - 1.12.x
   - 1.13.x
+  - 1.14.x
+  - 1.15.x
   - tip
 
 matrix:
   allow_failures:
     - go: tip
-
-  exclude:
-    - os: osx
-      go: 1.2.x
-    - os: osx
-      go: 1.3.x
-    - os: osx
-      go: 1.4
-    - os: osx
-      go: 1.5.x
-    - os: osx
-      go: 1.6.x
 
 before_install:
   - if [[ $TRAVIS_GO_VERSION == 1.13* ]]; then go get github.com/axw/gocov/gocov github.com/modocache/gover github.com/mattn/goveralls; fi