--- a/dedent/dedent.go
+++ b/dedent/dedent.go
@@ -1,9 +1,6 @@
-// +build !go1.10
-
 package dedent
 
 import (
-	"bytes"
 	"strings"
 )
 
@@ -28,7 +25,7 @@ func String(s string) string {
 		return s
 	}
 
-	var buf bytes.Buffer
+	var buf strings.Builder
 	for _, l := range lines {
 		l = strings.TrimPrefix(l, strings.Repeat(" ", minIndent))
 		buf.WriteString(l + "\n")