--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,4 @@ module github.com/muesli/reflow
 
 go 1.13
 
-require github.com/mattn/go-runewidth v0.0.7
+require github.com/mattn/go-runewidth v0.0.9