This adds a `dedent` package. Closes #5.

Instead of the implementation provided in the issue, I use a `strings.Builder` for Go 1.10 and onwards, which was a very easy performance improvement:

```
name      old time/op    new time/op    delta
Dedent-8     835ns ± 3%     656ns ± 8%  -21.50%  (p=0.000 n=9+10)

name      old alloc/op   new alloc/op   delta
Dedent-8      344B ± 0%      328B ± 0%   -4.65%  (p=0.000 n=10+10)

name      old allocs/op  new allocs/op  delta
Dedent-8      11.0 ± 0%       9.0 ± 0%  -18.18%  (p=0.000 n=10+10)
```