@lukasmalkmus 2020-10-18
Hm, the build failure seems related to coveralls, not my code additions.