@lukasmalkmus 2020-10-18
Ups, got it. CI tests against very old go versions. I don't know your stance regarding backwards compatibility but I just made it work by using a `bytes.Buffer` for Go 1.9 and before.