@muesli 2020-10-18
@lukasmalkmus I think it's fine to drop Go 1.10 support in the next release. Go ahead, use the StringBuilder and just remove 1.10 from the CI config.