@lukasmalkmus 2020-10-18
Alright, done. Ditched Go 1.9 support and lower, updated CI to test 1.14 and 1.15 and updated the only dependency to something more recent.