@lukasmalkmus 2020-10-18
I just ran `go mod tidy`. Maybe it is a good idea to squash before merging.