@lukasmalkmus 2020-10-18
Awesome, thanks for you quick response. Do you mind cutting a new release, soon?

Cheers.