@muesli 2020-10-18
No, thank you for the contribution! I'll release a `v0.2.0` tonight!